requirements.txt
flask
python-dotenv
gunicorn>=21.2.0
pyyaml
langsmith
langchain_anthropic
langchain_openai
langchain_community
langchain_core
langgraph
sqlalchemy
psycopg2-binary 